Extract text input into a component
